+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Short filter (`short_nav`) for nav items in page mode ([#41](https://github.com/daizutabi/mkapi/issues/41)).
 
+### Changed
+- Packages/modules without docstring are processed if their members have docstring. ([#42](https://github.com/daizutabi/mkapi/issues/42))
+
 ## [1.0.13] - 2020-08-03
 ### Fixed
 - Handle parsing empty line ([#23](https://github.com/daizutabi/mkapi/pull/23)). Thanks to [tony](https://github.com/tony).